Add Store&ItemServiceInterface
